Provide a way to make callbacks work with the cancel element
When cancel is selected, there is no corresponding input element, which means
that this is set to null in the callback function.

The "link" parameter to the callback function is set to the first 'a' child of
null, which translates to the first link on the page.

Fix the link parameter, by setting a rater-cancel-# (where # is the rater
serial number) class on the cancel element, so that we can identify the element
later in the select function.