google code issue 144
I needed the star to inherit the data collection from the radio input.
I figure this is a pretty cool thing to do and I tried it, it works great, it makes the plugin better, and now I suggest that you could make the same change to the master copy.
If this was on something like github then I would fork it and push it and send you a pull request.